chore: rename Authorized to S3Invocation
Move from authorized.rs to s3/invocation.rs. The type represents an
S3 request that has been authorized (has a presigned URL) and is ready
for HTTP execution.
