+++ rust/dialog-remote-s3/src/s3/provider.rs
@@ -1,8 +1,8 @@
 //! Provider implementations for S3 sites.
 //!
 //! Each module provides two layers:
-//! - `Provider<Authorized<Fx>>` on `Http` — shared HTTP execution (used by S3 and UCAN)
-//! - `Provider<Fork<S3, Fx>>` on `S3` — SigV4 authorization, then delegates to `Http`
+//! - `Provider<S3Invocation<Fx>>` — HTTP execution of authorized requests
+//! - `Provider<ForkInvocation<S3, Fx>>` — authorization via S3Authorization, then delegation
 
 pub mod archive;
 pub mod memory;
