+++ rust/dialog-remote-s3/src/s3/provider/archive.rs
@@ -1,15 +1,14 @@
 //! Archive capability providers for S3.
 //!
 //! Each effect is paired: `Provider<ForkInvocation<S3, Fx>>` authorizes via SigV4,
-//! then delegates to `Provider<Authorized<Fx>>` for HTTP execution.
+//! then delegates to `Provider<S3Invocation<Fx>>` for HTTP execution.
 
 use async_trait::async_trait;
 use dialog_capability::fork::ForkInvocation;
 use dialog_capability::{Policy, Provider};
 use dialog_effects::archive::*;
 
-use crate::Authorized;
-use crate::s3::{RequestDescriptorExt, S3};
+use crate::s3::{RequestDescriptorExt, S3, S3Invocation};
 
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
@@ -24,7 +23,7 @@ impl Provider<ForkInvocation<S3, Get>> for S3 {
             .await
             .map_err(|e| ArchiveError::Io(e.to_string()))?;
 
-        Authorized::new(permit, invocation.capability)
+        S3Invocation::new(permit, invocation.capability)
             .perform(self)
             .await
     }
@@ -32,8 +31,8 @@ impl Provider<ForkInvocation<S3, Get>> for S3 {
 
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-impl Provider<Authorized<Get>> for S3 {
-    async fn execute(&self, input: Authorized<Get>) -> Result<Option<Vec<u8>>, ArchiveError> {
+impl Provider<S3Invocation<Get>> for S3 {
+    async fn execute(&self, input: S3Invocation<Get>) -> Result<Option<Vec<u8>>, ArchiveError> {
         let client = reqwest::Client::new();
         let response = input
             .permit
@@ -69,7 +68,7 @@ impl Provider<ForkInvocation<S3, Put>> for S3 {
             .await
             .map_err(|e| ArchiveError::Io(e.to_string()))?;
 
-        Authorized::new(permit, invocation.capability)
+        S3Invocation::new(permit, invocation.capability)
             .perform(self)
             .await
     }
@@ -77,8 +76,8 @@ impl Provider<ForkInvocation<S3, Put>> for S3 {
 
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-impl Provider<Authorized<Put>> for S3 {
-    async fn execute(&self, input: Authorized<Put>) -> Result<(), ArchiveError> {
+impl Provider<S3Invocation<Put>> for S3 {
+    async fn execute(&self, input: S3Invocation<Put>) -> Result<(), ArchiveError> {
         let content = Put::of(&input.capability).content.clone();
 
         let client = reqwest::Client::new();
